feat: extract scrollable components (#2)
* chore: added redash

* chore: extract flatlist & scrollview from scrollbottomsheet

* chore: added sectionlist
