+++ example/src/screens/Home.tsx
@@ -1,69 +1,75 @@
-import * as React from 'react';
-import { View, Text, StyleSheet, Dimensions } from 'react-native';
+import React, { useCallback } from 'react';
+import {
+  View,
+  Text,
+  StyleSheet,
+  TouchableHighlight,
+  StatusBar,
+} from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
 import { StackNavigationProp } from '@react-navigation/stack';
-import ScrollBottomSheet from 'react-native-scroll-bottom-sheet';
 import { HomeStackParamsList } from '../App';
 
 interface HomeProps {
   navigation: StackNavigationProp<HomeStackParamsList, 'Home'>;
 }
 
-const windowHeight = Dimensions.get('window').height;
-
 const Home = ({ navigation }: HomeProps) => {
+  const handleFlatListPress = useCallback(() => {
+    navigation.navigate('FlatListExample');
+  }, [navigation]);
+
+  const handleSectionListPress = useCallback(() => {
+    navigation.navigate('SectionListExample');
+  }, [navigation]);
+
+  const handleScrollViewPress = useCallback(() => {
+    navigation.navigate('ScrollViewExample');
+  }, [navigation]);
   return (
-    <View style={styles.container}>
-      <ScrollBottomSheet<string> // If you are using TS, that'll infer the renderItem `item` type
-        componentType="FlatList"
-        snapPoints={[128, '50%', windowHeight - 200]}
-        initialSnapIndex={2}
-        renderHandle={() => (
-          <View style={styles.header}>
-            <View style={styles.panelHandle} />
-          </View>
-        )}
-        data={Array.from({ length: 20 }).map((_, i) => String(i))}
-        keyExtractor={i => i}
-        renderItem={({ item }) => (
-          <View style={styles.item}>
-            <Text>{`Item ${item}`}</Text>
-          </View>
-        )}
-        contentContainerStyle={styles.contentContainerStyle}
-      />
-    </View>
+    <SafeAreaView style={styles.root}>
+      <StatusBar barStyle="light-content" />
+      <View style={styles.container}>
+        <TouchableHighlight
+          style={styles.buttonContainer}
+          onPress={handleFlatListPress}
+        >
+          <Text>FlatList Example</Text>
+        </TouchableHighlight>
+        <TouchableHighlight
+          style={styles.buttonContainer}
+          onPress={handleSectionListPress}
+        >
+          <Text>SectionList Example</Text>
+        </TouchableHighlight>
+        <TouchableHighlight
+          style={styles.buttonContainer}
+          onPress={handleScrollViewPress}
+        >
+          <Text>ScrollView Example</Text>
+        </TouchableHighlight>
+      </View>
+    </SafeAreaView>
   );
 };
 
 const styles = StyleSheet.create({
-  container: {
+  root: {
     flex: 1,
     backgroundColor: 'black',
   },
-
-  contentContainerStyle: {
-    padding: 16,
-    backgroundColor: '#F3F4F9',
-  },
-  header: {
+  container: {
+    flex: 1,
+    backgroundColor: 'black',
+    alignContent: 'center',
     alignItems: 'center',
-    backgroundColor: 'white',
-    paddingVertical: 20,
-    borderTopLeftRadius: 20,
-    borderTopRightRadius: 20,
-  },
-  panelHandle: {
-    width: 40,
-    height: 2,
-    backgroundColor: 'rgba(0,0,0,0.3)',
-    borderRadius: 4,
-  },
-  item: {
-    padding: 20,
     justifyContent: 'center',
+  },
+  buttonContainer: {
+    padding: 12,
+    marginBottom: 12,
+    borderRadius: 6,
     backgroundColor: 'white',
-    alignItems: 'center',
-    marginVertical: 10,
   },
 });
