+++ src/hooks.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ScrollBottomSheetContext } from './context';
+
+export const useScrollBottomSheet = () => {
+  return useContext(ScrollBottomSheetContext);
+};
